chore: delay some stuff
