+++ blocks/ads/ads.js
@@ -96,62 +96,63 @@ export default function decorate(block) {
   const config = readBlockConfig(block);
   block.innerHTML = '';
 
-  const intersectHandler = () => {
-    window.pgatour = window.pgatour || {};
-
-    loadScript('/blocks/ads/jquery-3.6.0.min.js', () => {
-      loadScript('/blocks/ads/react-cq.min.js', () => {
-        loadScript('/blocks/ads/pgatour.min.js', () => {
-          // setup ads
-          window.pgatour.EngageTimer.setup();
-          window.pgatour.Ad.setup({
-            site: 'pgat',
-            refreshDisabled: false,
-            trackBrowserActivity: true,
-            justInTime: true,
-            refreshOnScroll: 'none',
-            useEngageTime: true,
-            options: {
-              s1: 'pgatour',
-              s2: 'tournaments',
-              s3: 'the-players',
-              s4: 'landing',
-            },
-            enableSingleRequest: true,
-            networkCode: '9517547',
-            refreshInterval: 20,
-          });
-          // setup ad wrapper
-          let wrapper;
-          let { position } = config;
-          switch (position) {
-            case 'leftpromo clock':
-              wrapper = buildLeftPromoClockEl(block);
-              position = 'leftpromo';
-              break;
-            case 'leftpromo toggle':
-              wrapper = buildLeftPromoToggleEl(block);
-              position = 'leftpromo';
-              break;
-            default:
-              break;
-          }
-          // create new ad
-          // eslint-disable-next-line no-new
-          new window.pgatour.Ad(wrapper, {
-            trackBrowserActivity: true,
-            options: { pos: position },
-            refreshOnResize: false,
-            companionAd: false,
-            justOnScroll: false,
-            suspended: false,
-            size: getAdSize(config.position),
+  const observer = new IntersectionObserver((entries) => {
+    if (entries.some((entry) => entry.isIntersecting)) {
+      observer.disconnect();
+      window.pgatour = window.pgatour || {};
+      loadScript('/blocks/ads/jquery-3.6.0.min.js', () => {
+        loadScript('/blocks/ads/react-cq.min.js', () => {
+          loadScript('/blocks/ads/pgatour.min.js', () => {
+            // setup ads
+            window.pgatour.EngageTimer.setup();
+            window.pgatour.Ad.setup({
+              site: 'pgat',
+              refreshDisabled: false,
+              trackBrowserActivity: true,
+              justInTime: true,
+              refreshOnScroll: 'none',
+              useEngageTime: true,
+              options: {
+                s1: 'pgatour',
+                s2: 'tournaments',
+                s3: 'the-players',
+                s4: 'landing',
+              },
+              enableSingleRequest: true,
+              networkCode: '9517547',
+              refreshInterval: 20,
+            });
+            // setup ad wrapper
+            let wrapper;
+            let { position } = config;
+            switch (position) {
+              case 'leftpromo clock':
+                wrapper = buildLeftPromoClockEl(block);
+                position = 'leftpromo';
+                break;
+              case 'leftpromo toggle':
+                wrapper = buildLeftPromoToggleEl(block);
+                position = 'leftpromo';
+                break;
+              default:
+                break;
+            }
+            // create new ad
+            // eslint-disable-next-line no-new
+            new window.pgatour.Ad(wrapper, {
+              trackBrowserActivity: true,
+              options: { pos: position },
+              refreshOnResize: false,
+              companionAd: false,
+              justOnScroll: false,
+              suspended: false,
+              size: getAdSize(config.position),
+            });
           });
         });
       });
-    });
-  };
+    }
+  }, { threshold: 0 });
 
-  const observer = new IntersectionObserver(intersectHandler, { threshold: 0 });
   observer.observe(block);
 }
