+++ blocks/player-feature/player-feature.js
@@ -1,4 +1,9 @@
-import { buildBlock, decorateBlock, loadBlock } from '../../scripts/scripts.js';
+import {
+  buildBlock,
+  decorateBlock,
+  decorateButtons,
+  loadBlock,
+} from '../../scripts/scripts.js';
 
 function transformBackgroundImage(section) {
   const picture = section.querySelector('picture');
@@ -28,6 +33,22 @@ function buildVideoContent(section) {
 }
 
 export default async function decorate(block) {
+  const link = block.querySelector('a');
+  const source = link.getAttribute('href');
+  const resp = await fetch(`${source}.plain.html`);
+  if (resp.ok) {
+    const html = await resp.text();
+    const feature = document.createElement('div');
+    feature.innerHTML = html;
+    block.innerHTML = `<div>${feature.querySelector('div').outerHTML}</div>`;
+    const video = block.querySelector('.embed, .video');
+    decorateButtons(block);
+    if (video) {
+      decorateBlock(video);
+      await loadBlock(video);
+    }
+  }
+
   const sections = ['background', 'status', 'name', 'credits'];
   [...block.firstElementChild.firstElementChild.children].forEach((child, i) => {
     if (sections[i]) child.classList.add(`player-feature-${sections[i]}`);
